requirements.txt
Django==3.0.7
djangorestframework==3.11.0
django-cors-headers==3.1.0
django-environ==0.4.4
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
django_extensions==2.2.9
